+++ src/main/java/com/moayo/moayoeats/backend/global/jwt/TokenService.java
@@ -2,15 +2,12 @@ package com.moayo.moayoeats.backend.global.jwt;
 
 import static com.moayo.moayoeats.backend.global.jwt.JwtUtil.AUTHORIZATION_HEADER;
 import static com.moayo.moayoeats.backend.global.jwt.JwtUtil.REFRESH_TOKEN_HEADER;
-import static com.moayo.moayoeats.backend.global.jwt.JwtUtil.REFRESH_TOKEN_TIME;
 
 import io.jsonwebtoken.Claims;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
 import java.io.IOException;
-import java.util.concurrent.TimeUnit;
 import lombok.RequiredArgsConstructor;
-import org.springframework.data.redis.core.RedisTemplate;
 import org.springframework.stereotype.Service;
 import org.springframework.util.StringUtils;
 
@@ -18,7 +15,7 @@ import org.springframework.util.StringUtils;
 @Service
 public class TokenService {
 
-    private final RedisTemplate<String, String> redisTemplate;
+    private final TokenRedisService tokenRedisService;
     private final JwtUtil jwtUtil;
 
     public String relatedIssuanceOfTokens(
@@ -52,9 +49,9 @@ public class TokenService {
                     info = jwtUtil.getUserInfoFromToken(accessToken);
                 }
                 String email = info.getSubject();
-                if (getRefreshToken(email) == null) { // refreshToken 만료시
+                if (tokenRedisService.getRefreshToken(email) == null) { // refreshToken 만료시
                     String newRefreshToken = jwtUtil.createRefreshToken(email); // refreshToken 생성
-                    saveRefreshToken(email, newRefreshToken);
+                    tokenRedisService.saveRefreshToken(email, newRefreshToken);
                     jwtUtil.addJwtToCookie(newRefreshToken, res, REFRESH_TOKEN_HEADER);
                 }
             } catch (Exception e) {
@@ -83,36 +80,5 @@ public class TokenService {
             res.sendRedirect("/login"); // 두 토큰 모두 만료
         }
         return "";
-    }
-
-    // refresh token 저장
-    public void saveRefreshToken(String email, String refreshToken) {
-
-        redisTemplate.opsForValue()
-            .set(email, refreshToken, REFRESH_TOKEN_TIME, TimeUnit.MILLISECONDS);
-    }
-
-    // refresh token 조회
-    public String getRefreshToken(String email) {
-
-        return redisTemplate.opsForValue().get(email);
-    }
-
-    // refresh token 삭제
-    public void deleteRefreshToken(String email) {
-
-        redisTemplate.opsForValue().getAndDelete(email);
-    }
-
-    // access token 블랙리스트 추가
-    public void setBlacklist(String accessToken, String message, Long expiration) {
-
-        redisTemplate.opsForValue().set(accessToken, message, expiration, TimeUnit.MILLISECONDS);
-    }
-
-    // 해당 access token이 블랙리스트에 있는지 확인
-    public boolean isBlacklisted(String accessToken) {
-
-        return redisTemplate.hasKey(accessToken);
     }
 }
